requirements.txt
tensorflow
numpy
scikit-learn
matplotlib
seaborn
pillow
pandas